type inference: use join for multiple predecessors
Previously an instruction with multiple predecessors would call its
transfer function once for each predecessor, and possibly alter the bags
for its successor arcs each time. This does not make sense. An example
of invalid inference:

(loop with x = (the fixnum y)
      do (if (consp x) (return nil) (setq x (the cons z))))

Previously this would infer x to be a fixnum throughout the loop,
because it happened to transfer that predecessor last. Needless to say
this would not end well.

Added an exported function INSTRUCTION-INPUT to make the join easy.