rewrite array and box handling in Cleavir
* delete the million aref/aset primops/asts/instructions
* replace with two: AREF, ASET, which store simplicity/type in slots
* delete the separate float box/unbox instructions
* replace with two: box- and unbox-, which store type in a slot

The AREF and ASET primops and ASTs have parameters/slots for whether the
things they're dealing with are boxed. AST-to-HIR looks at this
information and inserts box/unbox instructions conditionally.

These changes give implementors more latitude in choosing UAETs and
formats and bla bla bla, and also it lets there be like a million lines
in Cleavir be deleted.