Kildall escape and types: only track live variables
This change makes liveness define an extension class that filters pools
to only contain live variables. This implies some measure of defaulting
in find-in-pool. I made find-in-pool generic, but it might be better to
define an object-1 that receives a specialization and key (as is
necessary for type inference) instead. Not sure.

Provisional tests give this a speedup of seven times, which is good
because my test function is only a page long and it shouldn't take a
half a second just to infer types.

Further optimizations might be specific stuff like type caching
blablabla, but I suspect there would be greater improvements from using
something smarter than alists for pools.