Kildall: Rewrite to reduce consing
I'm not totally into the new internal interface but it works and it's
reasonably clean.

The basic idea here is that instead of following the paper slavishly and
making new pools for every transfer invocation, the old pool is directly
modified. This reduces consing to be one pool per instruction plus
whatever comparatively little consing transfer functions do to compute.
In some quick tests this reduced time and memory usage by half.