Revert "cleavir segregate-lexicals: Allocate cells as late as possible"
This reverts commit d5bff312a0d7df537ac95d37dae93b68c816bf57.

Clasp seems to have a write to an uninitialized cell somewhere with
this commit, and it seems likely to be flawed. map-instructions doesn't
handle control flow in quite the right way.

beach has another plan for fixing this involving changing how LET and
so on are compiled, which will avoid this problem.