Generate-AST: introduce and use primop for multiple-value-call
cl:multiple-value-call takes a function designator. Our multiple-value-call
AST and instruction only take functions, and have undefined behavior if this
is not the case. cl:funcall is in the same situation, so we do the same thing:
introduce a primop and assume that let implementations expand the CL operator
in its terms. Of course, m-v-c is a special operator rather than a function,
but the principle is identical.

This fixes an issue where use of multiple-value-call with non-functions would
cause undefined behavior.