Partial inlining: Do not copy lexical variable mapping
Though (I think) described in the paper, this results in bugs when
a variable is first defined in two branches which then join, as can
erise from e.g. function return values.

I'm not sure why the copying was done in the first place, but if it
is a good idea for some other reason, alternate solutions include:

 * A complex reconciliation step executed when a join is reached
 * Rewriting AST-to-HIR to not generate such code