Delete unused instructions
no-return is better unified with unreachable, i.e. if a call is
known to never return just have it proceed into an unreachable.
We don't have a tail call analysis but when we do we can
probably rather just put a flag on funcall-instruction to mark it
as being in tail position; for the purpose of lisp semantics,
being a tail call is irrelevant, so having it proceed into a
return instruction is fine.