Simplify function translation a bunch and make it more efficient.
We take advantage of the fact that function arguments can directly be
translated in the BIR argument class. This means we don't need to rely
on the function itself to bind any lexical variables as arguments and
can just receive the arguments directly. This has the added benefit
that there are less temporary lexical variables generated.