Change LOAD-TIME-VALUE representation.
Insert an explicit instruction for the load-time-value reference,
since now that it is explicitly for dumping, it logically must fetch
the value at load time from a constants table. Also keep track of the
load time values in the module to make translation simple. In the
future, we should move this out of BIR to allow LOAD-TIME-VALUE
representations to be client agnostic, as hoisting could happen
anywhere.