More reliably propagate types across function boundaries.
By hoisting the dumb "do-again" loop to the module-level, we can more
reliably propagate tpyes through LETI to other functions.

For example, consider (lambda (x)
                        (let ((y (the cons x)))
                          (lambda () (car y)))).

We'd like CAR to fold the CONS type test by seeing that Y is bound to
something asserted as CONS type, and by trying to fixpoint over the
module instead of individual functions, we have a better chance at
doing so.