Add :EXTERNAL THEs.
These THEs don't do any typechecking of their own, but merely assert
that the input value should be a certain type at the control point of
the THEI, and that someone else will do the checking. The compiler
should not assume the type has been proven at that point in time.