Remove DELETEDP slot from IBLOCK.
We no longer need it since now we maintain a flow order doubly linked
list. To check that we do not leave a deleted iblock in the flow order
we can NIL out the FUNCTION slot of a deleted iblock and use that to
verify that we don't keep around deleted iblocks.