Fix compilation error relating to blocks
Without this, the code assumes that anything before a prospective
if-if is a regular jump. This does not seem to be the case, e.g.
(lambda (stack)
  (not
   (block nil
     (labels ((f () (mapc (lambda (x) (return x)) stack)))
       (f)))))
on clasp (where mapc is marked dxcall) results in a previous
instruction being the block's catch.

I think this fix is correct, although maybe the if-if would still
be eliminable in this circumstance?