Fix (kludgily?) a meta evaluation bug
If meta evaluation of a function results in the deletion of
another function, as can happen e.g. if a checked THEI is deleted,
the meta evaluator could still try to meta evaluate it in the next
iteration. Given the function is trash at this point, this can end
poorly.

Not sure I like this fix. Editing a set/sequence as we go is
inherently dicey, but kind of baked in by now. Graph modifications
may require serious rethinking.