Rewrite attributes
The idea here is to make attributes something that can be
propagated kind of like types, but without worrying about some of
types' baggage like uncertain subtypep, or how to deal with
declarations. This should be better than the previous system, in
which attribute information only exists for very literal calls -
even (cleavir-primop:funcall #'foo ...) is too involved for the
previous system.