Fix common-return-cont
It would just take the first dynenv it saw to be common, and
functionally ignore any different ones. See
clasp-developers/clasp#1216, which seems to have been caused by
this issue.

Probably this whole thing should be rethought. Really, it should
be okay to contify even if the dynamic environments are different.
In the first case, if the function being contified returns, to be
contifiable all calls must end up at the same block anyway, and we
should be able to just locally unwind to that block's dynamic
environment before the call. In the second case, when the function
never returns, in correct code the ultimate destination of the
function is a parent of the dynenv of any call, so unwinding to
a shared ancestor or even farther shouldn't pose an issue.