Remove consistency assertions in BIR modifiers
These have mostly been an annoyance. While modifying the IR it is
in an inconsistent state, so these assertions trigger unless the
modifications take place in certain ways. And these certain ways
don't actually preserve consistency in the moment, so this only
serves to make the code less clear.

I think the basic rule should be that IR only needs to be
consistent at some points - between modifications - and that this
can be enforced by calling VERIFY appropriately.