Fix reversed subtypep in meta-evaluate
Thinking about this is slightly mind-bending, but I think it's
correct now. If the asserted type is a subtype of the check type,
for the assertion to be true the check must also be true, so we
can validly move up the check.

I haven't quite thought through when the intersection of the
asserted and derived types could be used here yet.