Abstract interpreter: refactor
This is a rewrite of the abstract interpreter. The goal was to
reduce the amount of boilerplate code each domain needs, as well
as to improve modularity.

First, domains now do not concern themselves with side effects.
They just specialize FLOW-INSTRUCTION which can be a pure function,
and the general abstract interpretation machinery takes care of
storing that information, etc.

Second, products have been rewritten to use "channels", roughly
like Cousot's. Domain definitions can now be self contained, while
interactions are put in channels instead.

This is still quite experimental. The obvious disadvantage with
the rewrite is it probably conses more, but it should just be
transient consing, so we'll see if it matters. I have only tested
the interpreter with very basic code so we'll see if it holds up.