Abstract interpreter: Expand initialization phase
This makes the interpreter more self-contained, in that it doesn't
have to coordinate so much with the initforms of slots and stuff.
Also important because the initial state for reachablep is slightly
complicated. And because I do want to try optimistic analyses.

Downside is it takes some more time. Blech.