CST-to-AST: Consolidate TYPE-WRAP
Having multiple hook functions that do almost the same thing is
an annoying organization.

The extra contexts are put in due to my experience with Clasp. If
the type checker has to return the values it's given (as we
require; might be worth reviewing again) and we don't know how
many values are coming in as from almost all Lisp type specifiers,
the type checker has to do an expensive (apply #'values ...)
operation which is hard to optimize away in all situations.
Catching the common situations upfront within CST-to-AST
substantially reduces the burden on the optimizer.

Also, different contexts makes it easy to have different policies
depending on the source of the type declaration.