Represent function+variable cells in BIR
Dr. Strandh described cells quite a while ago but used
LOAD-TIME-VALUE for them. This works, but can add to the amount of
code that needs processing. FASLs can be a lot more efficient if
they have specific operations for linking these cells. And it
makes processing of various kinds easier and simpler.