BIR transformations: axe code subsequent to nonreturning function
Can happen if a type deriver figures a function never returns.
The most obvious case is when there's a call to ERROR that the
frontend hasn't taken care of, which could include e.g.
(apply #'error ...), but it could also come from figuring a call
is getting a wrongly typed argument. Though in that case we should
probably raise a warning.